+++ zam/net/acceptor/acceptorTcp.h
@@ -8,8 +8,6 @@
 #include <zam/base/exception.h>
 #include "acceptor.h"
 
-#include "../connection/connectionTcp.h"
-
 #include <boost/bind.hpp>
 
 namespace zam {
@@ -24,66 +22,14 @@ namespace zam {
                 };
 
             public:
-                explicit acceptorTcp(base::io::ioSystem &ios, warehouse::warehouse w, Config cfg)
-                        : acceptor(ios, std::move(w))
-                        , cfg_(std::move(cfg))
-                        , acceptorSocket_(ios.getIos())
-                {}
+                ZAMNET_API explicit acceptorTcp(base::io::ioSystem &ios, warehouse::warehouse w, Config cfg);
 
-                void startAccept() override {
-                    using tcp = boost::asio::ip::tcp;
-
-                    tcp::resolver resolver(ios().getIos());
-
-                    std::string ipAddr(cfg_.bindIp);
-                    if (ipAddr.empty()) {
-                        ipAddr = "0.0.0.0";
-                    }
-                    auto port = std::to_string(cfg_.bindPort);
-
-                    boost::system::error_code ec;
-
-                    tcp::resolver::query query(ipAddr, port);
-                    tcp::endpoint endpoint = *resolver.resolve(query, ec);
-
-                    acceptorSocket_.open(endpoint.protocol(), ec);
-
-                    acceptorSocket_.set_option(tcp::no_delay(true));
-                    acceptorSocket_.set_option(tcp::acceptor::reuse_address(true));
-                    acceptorSocket_.set_option(boost::asio::socket_base::linger(true, 0));
-                    acceptorSocket_.bind(endpoint, ec);
-
-                    if (!!ec) {
-                        throw base::zamException(netError::failed_bind, ec.message());
-                    }
-
-                    acceptorSocket_.listen(boost::asio::socket_base::max_connections, ec);
-                    if (!!ec) {
-                        throw base::zamException(netError::failed_listen, ec.message());
-                    }
-
-                    startAcceptInternal();
-                }
+                ZAMNET_API void startAccept() override;
 
             private:
-                void acceptHandler(const boost::system::error_code &ec) {
-                    if (!ec) {
-                        conn_->startAccept();
+                void acceptHandler(const boost::system::error_code &ec);
 
-                        conn_->ioPost(boost::bind(&handler::eventHandler::onAccept, conn_->eventHandler(), conn_));
-
-                        startAcceptInternal();
-                    }
-                }
-
-                void startAcceptInternal() {
-                    auto c = new connection::connectionTcp(ios());
-                    c->initialize(this->warehouse());
-                    conn_.reset(c);
-
-                    acceptorSocket_.async_accept(c->socket(), strand().wrap(
-                            boost::bind(&acceptorTcp::acceptHandler, this, boost::asio::placeholders::error)));
-                }
+                void startAcceptInternal();
 
             private:
                 Config cfg_;
