+++ zam/net/connector/connector.h
@@ -1,0 +1,21 @@
+//
+// Created by SungWoo Yang on 2017. 9. 21..
+//
+
+#ifndef ZAMLIB_CONNECTOR_H
+#define ZAMLIB_CONNECTOR_H
+
+namespace zam {
+    namespace net {
+        namespace connector {
+
+            struct connector
+            {
+
+            };
+
+        }
+    }
+}
+
+#endif //ZAMLIB_CONNECTOR_H
