+++ zam/net/connector/connectorTcp.h
@@ -1,0 +1,22 @@
+//
+// Created by SungWoo Yang on 2017. 9. 21..
+//
+
+#ifndef ZAMLIB_CONNECTORTCP_H
+#define ZAMLIB_CONNECTORTCP_H
+
+#include "connector.h"
+
+namespace zam {
+    namespace net {
+        namespace connector {
+
+            class connectorTcp {
+
+            };
+
+        }
+    }
+}
+
+#endif //ZAMLIB_CONNECTORTCP_H
